+++ command/error_recorder.go
@@ -2,8 +2,7 @@ package command
 
 import (
 	"context"
-
-	"github.com/pkg/errors"
+	"fmt"
 
 	"github.com/get-eventually/go-eventually"
 	"github.com/get-eventually/go-eventually/event"
@@ -27,13 +26,13 @@ type ErrorRecorder struct {
 	// Appender is the Event Store instance used for appending domain events.
 	Appender event.Appender
 
-	// CaptureErrors specifies whether an error reported from the
-	// command.Handler should be captured (or "silenced") by this component
-	// and return nil instead.
+	// ShouldCaptureError is a function that should specify
+	// whether an error reported from the command.Handler should be captured (or "silenced")
+	// by this component and avoid returning it to the command caller.
 	//
 	// Default behavior is to return all errors from the command.Handler
 	// to the caller.
-	CaptureErrors bool
+	ShouldCaptureError func(err error) bool
 
 	// StreamType specifies the stream type used for the events produced by this component.
 	// If unspecified, FailedCommandType will be used by default.
@@ -76,10 +75,10 @@ func (er ErrorRecorder) buildStreamID(cmd Command) event.StreamID {
 // in case of failure, appends the error and command to the Event Store
 // using the user-provided EventMapper.
 //
-// If CaptureError has been set to "false", the error coming from the command.Handler
+// If ShouldCaptureError has been set and returns "false", the error coming from the command.Handler
 // will be returned to the caller.
 //
-// If CaptureError has been set to "true", the error from the command.Handler is silenced
+// If ShouldCaptureError has been set and returns "true", the error from the command.Handler is silenced
 // but an error can still be returned if the append operation on the Event Store fails.
 func (er ErrorRecorder) Handle(ctx context.Context, cmd Command) error {
 	err := er.Handler.Handle(ctx, cmd)
@@ -92,14 +91,19 @@ func (er ErrorRecorder) Handle(ctx context.Context, cmd Command) error {
 		Payload: er.EventMapper(err, cmd),
 	}
 
-	_, appendErr := er.Appender.Append(ctx, streamID, version.Any, event)
-	if appendErr != nil && er.CaptureErrors {
-		// Append error only returned if silencing command.Handler errors.
-		return errors.Wrap(err, "command.ErrorRecorder: failed to append command error to event store")
+	captureError := false
+	if er.ShouldCaptureError != nil {
+		captureError = er.ShouldCaptureError(err)
 	}
 
-	if !er.CaptureErrors {
-		return errors.Wrap(err, "command.ErrorRecorder: command handler failed")
+	_, appendErr := er.Appender.Append(ctx, streamID, version.Any, event)
+	if appendErr != nil && captureError {
+		// Append error only returned if silencing command.Handler errors.
+		return fmt.Errorf("command.ErrorRecorder: failed to append command error to event store: %w", err)
+	}
+
+	if !captureError {
+		return fmt.Errorf("command.ErrorRecorder: command handler failed: %w", err)
 	}
 
 	return nil