+++ command/error_recorder_test.go
@@ -98,7 +98,7 @@ func TestErrorRecorder(t *testing.T) {
 		}, trackingEventStore.Recorded())
 	})
 
-	t.Run("when handler fails and CaptureError is set to true, no error is returned", func(t *testing.T) {
+	t.Run("when handler fails and ShouldCaptureError returns true, no error is returned", func(t *testing.T) {
 		eventStore := inmemory.NewEventStore()
 		trackingEventStore := inmemory.NewTrackingEventStore(eventStore)
 
@@ -108,11 +108,13 @@ func TestErrorRecorder(t *testing.T) {
 		}
 
 		handler := command.ErrorRecorder{
-			CaptureErrors: true,
 			Handler: command.HandlerFunc(func(ctx context.Context, cmd command.Command) error {
 				return expectedErr
 			}),
 			Appender: trackingEventStore,
+			ShouldCaptureError: func(err error) bool {
+				return true
+			},
 			EventMapper: func(err error, cmd command.Command) eventually.Payload {
 				return mockCommandHasFailed{
 					err:     err,
@@ -124,6 +126,53 @@ func TestErrorRecorder(t *testing.T) {
 		err := handler.Handle(context.Background(), expectedCommand)
 
 		assert.NoError(t, err)
+		assert.Equal(t, []event.Persisted{
+			{
+				Version: 1,
+				Stream: event.StreamID{
+					Type: command.FailedType,
+					Name: expectedCommand.Payload.Name(),
+				},
+				Event: event.Event{
+					Payload: mockCommandHasFailed{
+						err:     expectedErr,
+						command: expectedCommand.Payload.(mockCommand),
+					},
+				},
+			},
+		}, trackingEventStore.Recorded())
+	})
+
+	t.Run("when handler fails and ShouldCaptureError returns false, error is returned to caller", func(t *testing.T) {
+		eventStore := inmemory.NewEventStore()
+		trackingEventStore := inmemory.NewTrackingEventStore(eventStore)
+
+		expectedErr := errors.New("failed command")
+		unexpectedErr := errors.New("unexpected error")
+
+		expectedCommand := command.Command{
+			Payload: mockCommand{message: t.Name()},
+		}
+
+		handler := command.ErrorRecorder{
+			Handler: command.HandlerFunc(func(ctx context.Context, cmd command.Command) error {
+				return expectedErr
+			}),
+			Appender: trackingEventStore,
+			ShouldCaptureError: func(err error) bool {
+				return errors.Is(err, unexpectedErr) // This will return false
+			},
+			EventMapper: func(err error, cmd command.Command) eventually.Payload {
+				return mockCommandHasFailed{
+					err:     err,
+					command: cmd.Payload.(mockCommand),
+				}
+			},
+		}
+
+		err := handler.Handle(context.Background(), expectedCommand)
+
+		assert.ErrorIs(t, err, expectedErr)
 		assert.Equal(t, []event.Persisted{
 			{
 				Version: 1,