+++ extension/postgres/row_streamer.go
@@ -6,11 +6,11 @@ import (
 	"fmt"
 
 	"github.com/get-eventually/go-eventually"
-	"github.com/get-eventually/go-eventually/eventstore"
+	"github.com/get-eventually/go-eventually/event"
 	"github.com/get-eventually/go-eventually/logger"
 )
 
-func rowsToStream(rows *sql.Rows, es eventstore.EventStream, r eventstore.Registry, l logger.Logger) error {
+func rowsToStream(rows *sql.Rows, es event.Stream, deserializer Deserializer, l logger.Logger) error {
 	defer func() {
 		if err := rows.Close(); err != nil {
 			logger.Error(l, "Failed to close streamed event rows", logger.With("err", err))
@@ -20,16 +20,16 @@ func rowsToStream(rows *sql.Rows, es eventstore.EventStream, r eventstore.Regist
 	for rows.Next() {
 		var (
 			eventName               string
-			event                   eventstore.Event
+			evt                     event.Persisted
 			rawPayload, rawMetadata json.RawMessage
 		)
 
 		err := rows.Scan(
-			&event.SequenceNumber,
-			&event.Stream.Type,
-			&event.Stream.Name,
+			&evt.SequenceNumber,
+			&evt.Stream.Type,
+			&evt.Stream.Name,
 			&eventName,
-			&event.Version,
+			&evt.Version,
 			&rawPayload,
 			&rawMetadata,
 		)
@@ -37,7 +37,7 @@ func rowsToStream(rows *sql.Rows, es eventstore.EventStream, r eventstore.Regist
 			return fmt.Errorf("postgres.EventStore: failed to scan stream row into event struct: %w", err)
 		}
 
-		payload, err := r.Deserialize(eventName, rawPayload)
+		payload, err := deserializer.Deserialize(eventName, rawPayload)
 		if err != nil {
 			return fmt.Errorf("postgres.EventStore: failed to deserialize event: %w", err)
 		}