+++ CHANGELOG.md
@@ -7,9 +7,14 @@ project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 ### Added
-- ...
+- An option to override Event appending logic in Postgres EventStore implementation.
+- `postgres.Serde` interface to support more serialization formats.
 
 ### Changed
+- Existing `Event-Id` value in Event Metadata does not get overwritten in correlation.EventStoreWrapper.
+- `postgres.EventStore` now uses the `Serde` interface for serializing to and deserializing from byte array.
+- `postgres.Registry` is now called `postgres.JSONRegistry` and implements thenew `postgres.Serde` interface.
+- `CaptureErrors` in `command.ErrorRecorder` is now a function (`ShouldCaptureError`), to allow for a more flexible capture strategy.
 - `projection` package has been removed, the types have been moved:
     - `projection.Applier` is now `event.Processor`,
     - `projection.Runner` is now `event.ProcessorRunner`.